# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCADA Logistics API",
    "description": "Мониторинг логистической инфраструктуры предприятия.\n\n**Авторизация:** `POST /auth/login` → `Authorization: Bearer <access_token>`; access живёт 15 мин, обновление — `POST /auth/refresh`. WebSocket: `/ws/live?token=<access_token>`. Демо-пользователи: `dispatcher`, `security`, `admin`, пароль `demo`.\n\nКоординаты — метры локального плана (см. `GET /objects`). Время — ISO 8601 с таймзоной.\n\nWebSocket `/ws/live`: формат сообщений — `contracts/ws.schema.json`.",
    "version": "0.2.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Wrong username or password"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "New token pair; role and permissions are re-read, so role changes apply on the next refresh.",
        "operationId": "refresh_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/kpi": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Kpi",
        "description": "Показатели за период (по умолчанию — последний час).",
        "operationId": "kpi_kpi_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — час назад",
              "title": "From"
            },
            "description": "По умолчанию — час назад"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — сейчас",
              "title": "To"
            },
            "description": "По умолчанию — сейчас"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KpiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/heatmap": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Heatmap",
        "description": "Загруженность территории по ГЛОНАСС: сколько замеров попало в каждую ячейку сетки.",
        "operationId": "heatmap_heatmap_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — час назад",
              "title": "From"
            },
            "description": "По умолчанию — час назад"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — сейчас",
              "title": "To"
            },
            "description": "По умолчанию — сейчас"
          },
          {
            "name": "cell",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100,
              "minimum": 2,
              "description": "Размер ячейки, м",
              "default": 10,
              "title": "Cell"
            },
            "description": "Размер ячейки, м"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "vehicles",
                "stops"
              ],
              "type": "string",
              "description": "vehicles — все позиции; stops — где стоят",
              "default": "vehicles",
              "title": "Source"
            },
            "description": "vehicles — все позиции; stops — где стоят"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Heatmap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/replay": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Replay",
        "description": "Треки всех машин за окно (до 6 ч) с шагом `step` и тревоги этого окна — для таймлайна воспроизведения.",
        "operationId": "replay_replay_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — час назад",
              "title": "From"
            },
            "description": "По умолчанию — час назад"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "По умолчанию — сейчас",
              "title": "To"
            },
            "description": "По умолчанию — сейчас"
          },
          {
            "name": "step",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 600,
              "minimum": 1,
              "description": "Шаг выборки, с",
              "default": 5,
              "title": "Step"
            },
            "description": "Шаг выборки, с"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Replay"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertsKpi": {
        "properties": {
          "opened": {
            "type": "integer",
            "title": "Opened"
          },
          "by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Severity"
          },
          "top_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleCount"
            },
            "type": "array",
            "title": "Top Rules"
          },
          "mean_ack_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Ack S",
            "description": "Среднее время от открытия до подтверждения"
          },
          "open_now": {
            "type": "integer",
            "title": "Open Now"
          }
        },
        "type": "object",
        "required": [
          "opened",
          "by_severity",
          "top_rules",
          "mean_ack_s",
          "open_now"
        ],
        "title": "AlertsKpi"
      },
...
      "BuildingTraffic": {
        "properties": {
          "building_id": {
            "type": "string",
            "title": "Building Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "entries": {
            "type": "integer",
            "title": "Entries",
            "description": "Проходов внутрь по СКУД"
          },
          "people_now": {
            "type": "integer",
            "title": "People Now"
          }
        },
        "type": "object",
        "required": [
          "building_id",
          "name",
          "entries",
          "people_now"
        ],
        "title": "BuildingTraffic"
      },
...
      "GateKpi": {
        "properties": {
          "entries": {
            "type": "integer",
            "title": "Entries",
            "description": "Въезды в зону КПП-1 по ГЛОНАСС"
          },
          "exits": {
            "type": "integer",
            "title": "Exits"
          },
          "plates_recognized": {
            "type": "integer",
            "title": "Plates Recognized",
            "description": "Распознаваний номеров камерами КПП"
          },
          "by_hour": {
            "items": {
              "$ref": "#/components/schemas/HourCount"
            },
            "type": "array",
            "title": "By Hour"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "exits",
          "plates_recognized",
          "by_hour"
        ],
        "title": "GateKpi"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeatCell": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X",
            "description": "Центр ячейки, м"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "count"
        ],
        "title": "HeatCell"
      },
      "Heatmap": {
        "properties": {
          "cell_m": {
            "type": "number",
            "title": "Cell M"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "max_count": {
            "type": "integer",
            "title": "Max Count"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/HeatCell"
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "type": "object",
        "required": [
          "cell_m",
          "source",
          "max_count",
          "cells"
        ],
        "title": "Heatmap"
      },
...
      "HourCount": {
        "properties": {
          "hour": {
            "type": "string",
            "format": "date-time",
            "title": "Hour"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "exits": {
            "type": "integer",
            "title": "Exits"
          }
        },
        "type": "object",
        "required": [
          "hour",
          "entries",
          "exits"
        ],
        "title": "HourCount"
      },
      "KpiResponse": {
        "properties": {
          "period_from": {
            "type": "string",
            "format": "date-time",
            "title": "Period From"
          },
          "period_to": {
            "type": "string",
            "format": "date-time",
            "title": "Period To"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/VehicleKpi"
            },
            "type": "array",
            "title": "Vehicles"
          },
          "gate": {
            "$ref": "#/components/schemas/GateKpi"
          },
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/BuildingTraffic"
            },
            "type": "array",
            "title": "Buildings"
          },
          "alerts": {
            "$ref": "#/components/schemas/AlertsKpi"
          }
        },
        "type": "object",
        "required": [
          "period_from",
          "period_to",
          "vehicles",
          "gate",
          "buildings",
          "alerts"
        ],
        "title": "KpiResponse"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "Replay": {
        "properties": {
          "period_from": {
            "type": "string",
            "format": "date-time",
            "title": "Period From"
          },
          "period_to": {
            "type": "string",
            "format": "date-time",
            "title": "Period To"
          },
          "step_s": {
            "type": "integer",
            "title": "Step S"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/ReplayTrack"
            },
            "type": "array",
            "title": "Tracks"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "type": "array",
            "title": "Alerts",
            "description": "Тревоги, открытые в этом окне: для отметок на таймлайне"
          }
        },
        "type": "object",
        "required": [
          "period_from",
          "period_to",
          "step_s",
          "tracks",
          "alerts"
        ],
        "title": "Replay"
      },
      "ReplayTrack": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/RoutePoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "points"
        ],
        "title": "ReplayTrack"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "dispatcher",
          "security",
          "admin"
        ],
        "title": "Role"
      },
...
      "RoutePoint": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "speed_kmh": {
            "type": "number",
            "title": "Speed Kmh"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "x",
          "y",
          "speed_kmh"
        ],
        "title": "RoutePoint"
      },
      "RuleCount": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "count"
        ],
        "title": "RuleCount"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Время жизни access-токена, секунд"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "TokenPair"
      },
      "User": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "examples": [
              "dispatcher"
            ]
          },
          "full_name": {
            "type": "string",
            "title": "Full Name",
            "examples": [
              "Иванов И. И."
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "full_name",
          "role"
        ],
        "title": "User"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VehicleKpi": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "plate": {
            "type": "string",
            "title": "Plate"
          },
          "mileage_km": {
            "type": "number",
            "title": "Mileage Km",
            "description": "По одометру трекера за период"
          },
          "moving_min": {
            "type": "number",
            "title": "Moving Min",
            "description": "Скорость > 1 км/ч"
          },
          "idle_min": {
            "type": "number",
            "title": "Idle Min",
            "description": "Стоит с работающим двигателем"
          },
          "stopped_min": {
            "type": "number",
            "title": "Stopped Min",
            "description": "Стоит с заглушенным двигателем"
          },
          "utilization_pct": {
            "type": "number",
            "title": "Utilization Pct",
            "description": "Доля времени в движении; всё считается только на территории"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "plate",
          "mileage_km",
          "moving_min",
          "idle_min",
          "stopped_min",
          "utilization_pct"
        ],
        "title": "VehicleKpi"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}